Move Plugin/xmlutils.* into CodeLite/ Replace duplicated helper functions by `XmlUtils` (`FindNodeByName`/`SetNodeContent`/`FindFirstByTagName`).
