+++ CodeLite/clEnvironment.cpp
@@ -3,8 +3,8 @@
 #include "StringUtils.h"
 #include "archive.h"
 #include "cl_standard_paths.h"
-#include "file_logger.h"
 #include "fileutils.h"
+#include "xmlutils.h"
 
 #include <wx/any.h>
 #include <wx/filename.h>
@@ -55,21 +55,6 @@ std::vector<std::pair<wxString, wxString>> FindVariablesInString(wxString str)
     return result;
 }
 
-wxXmlNode* FindFirstByTagName(const wxXmlNode* parent, const wxString& tagName)
-{
-    if(!parent) {
-        return nullptr;
-    }
-
-    wxXmlNode* child = parent->GetChildren();
-    while(child) {
-        if(child->GetName() == tagName) {
-            return child;
-        }
-        child = child->GetNext();
-    }
-    return nullptr;
-}
 } // namespace
 
 clEnvironment::clEnvironment()
@@ -80,7 +65,7 @@ clEnvironment::clEnvironment()
     if(!doc.Load(config.GetFullPath()))
         return;
 
-    wxXmlNode* node = FindFirstByTagName(doc.GetRoot(), "ArchiveObject");
+    wxXmlNode* node = XmlUtils::FindFirstByTagName(doc.GetRoot(), "ArchiveObject");
     if(node) {
         Archive arc;
         arc.SetXmlNode(node);
