+++ CodeLite/xmlutils.cpp
