+++ CodeLite/xmlutils.h
@@ -31,16 +31,16 @@
 #include <wx/xml/xml.h>
 
 /// A collection of XML utils
-class WXDLLIMPEXP_SDK XmlUtils
+class WXDLLIMPEXP_CL XmlUtils
 {
 public:
-    /// Find a child node by name by iterating the parent children. NULL if no childs exist
+    /// Find a child node by name by iterating the parent children. nullptr if no childs exist
     /// \param parent  the parent node whom to be searched
     /// \param tagName the element tag name
     /// \param name	   the element 'Name' property value to search
     static wxXmlNode* FindNodeByName(const wxXmlNode* parent, const wxString& tagName, const wxString& name);
 
-    /// Find the last child node of parent with a given name. NULL if no childs exist
+    /// Find the last child node of parent with a given name. nullptr if no childs exist
     /// \param parent  the parent node whom to be searched
     /// \param name	   the element's tag name
     static wxXmlNode* FindLastByTagName(const wxXmlNode* parent, const wxString& tagName);
@@ -55,7 +55,7 @@ public:
     static void UpdateProperty(wxXmlNode* node, const wxString& name, const wxString& value);
 
     /**
-     * Find the first child node of parent with a given name. NULL if no children exist
+     * Find the first child node of parent with a given name. nullptr if no children exist
      * \param the parent node to be searched
      * \param the element's tag name
      */
